--- Code/Moco/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Mon Apr 17 12:56:15 CEST 2023
-distributionBase=GRADLE_USER_HOME
-distributionUrl=https\://services.gradle.org/distributions/gradle-7.5-bin.zip
-distributionPath=wrapper/dists
-zipStorePath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
