+++ app/src/main/java/com/example/footprinthero/MainActivity.kt
@@ -1,0 +1,11 @@
+package com.example.footprinthero
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+}
